LPS-156903 Order the configuration items better Move Banner and Consent Panel configuration interfaces into packages to make Preference Handling always appear as first item in the sidebar.
