+++ modules/apps/cookies/cookies-api/src/main/resources/com/liferay/cookies/configuration/consent/packageinfo
@@ -1,0 +1,1 @@
+version 1.0.0
