+++ modules/apps/cookies/cookies-impl/src/main/java/com/liferay/cookies/internal/settings/definition/CookiesConsentConfigurationBeanDeclaration.java
@@ -14,7 +14,7 @@
 
 package com.liferay.cookies.internal.settings.definition;
 
-import com.liferay.cookies.configuration.CookiesConsentConfiguration;
+import com.liferay.cookies.configuration.consent.CookiesConsentConfiguration;
 import com.liferay.portal.kernel.settings.definition.ConfigurationBeanDeclaration;
 
 import org.osgi.service.component.annotations.Component;
